refactor(mesh-llm-ffi): split bindings by responsibility
